requirements.txt
asgiref==3.4.1
Django==3.2.7
django-autoslug==1.9.8
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-resized==0.3.11
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
pycep-correios==5.0.0